[IT-4485] Refactor to prep for adding new endpoint
Refactor the lambda into multiple submodules in preparation of adding
a new API endpoint for getting the current balances.

Also rename 'mips' to 'mip' to accurately reflect the name of the
upstream service.
