+++ template.yaml
@@ -87,7 +87,7 @@ Resources:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
       CodeUri: .
-      Handler: mips_api.lambda_handler
+      Handler: mip_api.lambda_handler
       Runtime: python3.12
       Environment:
         Variables:
